Change WorkItemStateMapping for work item sources from a selectable field to an interface implementation and update the tests. This way we can resolve sources for multiple work items sources at the project level more efficiently.
This also completely changes the old semantics of this field, since we are actually returning the work items source map now not the mapping
of the actual states in the work items. The latter is now more easily obtained by querying the unmapped states of the work items directly, which we are doing as part of the pipeline stats, so the old selectable field resolver is not very useful now.

[story=PO-91 subject=cycle_metrics_conditionals story_url=https://urjuna.atlassian.net/browse/PO-91]