In order to fix this bug, we made a change in a basic existing contract.
By default sub_tasks are now included in the cycle metrics calcs so that the numbers we see in the summary will match the tasks we see in the detail views.  This should actually work ok since now we are focusing on specs and if a subtask has commits it probably should show up in the cycle metrics too.

we also changed the include_epics_and_subtasks args in the cycle metric trends args api into two separate args, include_epics and include_sub_tasks that can be independently toggled. The defaults are include_epics=false and include_subtasks=true to reflect the new behaviors described above.

Updated tests to reflect the new contracts.

[story=PO-277  story_url=https://urjuna.atlassian.net/browse/PO-277]