Refactoring: move common utility functions for trending and calculating cycle metrics out of project/selectable so we can reuse them in team/selectable 1. Move basic date and timeline utilities in to top level graphql/utils.py 2. Move the CycleMetricTrends base class with metrics maps etc into work_items/sql_expressions. All the dependencies here are below work item and so this is the logical place for this to belong.
This leaves duplicated code in teams and projects around how the cycle metrics are calculated, but we can start figuring out how to lift that stuff (or whether we even should) after we have gone through the process of seeing how the requirements might differ between projects and teams. For now we will tolerate the code duplication.

[PO-740]